+++ spot_wrapper/calibration/spot_in_hand_camera_calibration.py
@@ -39,7 +39,7 @@ from bosdyn.client.robot_command import (
 from bosdyn.client.robot_state import RobotStateClient
 from bosdyn.client.time_sync import TimedOutError
 
-from spot_wrapper.calibration.automatic_camera_calibration_robot import (
+from spot_wrapper.calibration.calibration_helpers import (
     AutomaticCameraCalibrationRobot,
 )
 from spot_wrapper.calibration.charuco_board_detection import (
